test: Split some tests into integration tests
